Refactor database seed script
